+++ lib/k8s/sys/telemetry.ex
@@ -1,12 +1,22 @@
 defmodule K8s.Sys.Telemetry do
   @moduledoc false
 
-  @events [
-    [:http, :request, :start],
-    [:http, :request, :stop],
-    [:http, :request, :exception]
+  @spans [
+    [:http, :request]
   ]
 
+  @spec spans() :: list()
+  def spans, do: @spans
+
   @spec events() :: list()
-  def events, do: @events
+  def events do
+    @spans
+    |> Enum.flat_map(fn span ->
+      [
+        span ++ [:start],
+        span ++ [:stop],
+        span ++ [:exception]
+      ]
+    end)
+  end
 end